refactor: split oneshot/poll.rs into focused modules (#555)
Split 1824-line poll.rs into 6 modules under oneshot/poll/:
- mod.rs: public API (poll_messages, pull_messages, QueryOptions)
- meta.rs: read_host_from_meta, chat_path_from_meta
- filter_tier.rs: subscription tier filtering
- filter_events.rs: event type filtering
- commands.rs: cmd_poll, cmd_pull, cmd_watch, cmd_query, etc.
- multi_room.rs: poll_messages_multi

All 66 unit tests preserved in their respective modules.
No API changes — all public functions re-exported at same paths.

Closes #555
